Arreglos varios: - Mensajes de error y confirmación no funcionaban correctamente. - Tipos de datos en BD - Identación - Nueva lógica para la eliminación de archivos

+++ database/migrations/2016_05_05_142734_create_usuarios_table.php
@@ -15,7 +15,7 @@ class CreateUsuariosTable extends Migration {
 	{
 		Schema::create('usuarios', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('documento')->unsigned()->unique();
+            $table->bigInteger('documento')->unsigned()->unique();
             $table->string('apellido');
             $table->string('nombre');
             $table->timestamps();
